change token name and cntent and put register schema on file
